店舗検索処理実装中(一時保存)

+++ main-service/src/main/java/com/qms/mainservice/domain/model/valueobject/StoreMenuId.java
@@ -2,8 +2,8 @@ package com.qms.mainservice.domain.model.valueobject;
 
 import com.qms.shared.domain.model.BaseId;
 
-public record StoreMenuId(Long value) implements BaseId<Long> {
-    public static StoreMenuId of(Long value) {
+public record StoreMenuId(Integer value) implements BaseId<Integer> {
+    public static StoreMenuId of(Integer value) {
         return new StoreMenuId(value);
     }
 }
